Refactor logic with state machine and timers
Previously, I had haphazardly hacked the program together with much of
the logic being handled by the window and timerbutton classes.  This was
a bad idea on the whole, but it let me get a working release out
quickly.  However, the addition of a portrait mode requires more
timerbutton objects that should share their state with the landscape
ones, which couldn't be achieved the old way.

This commit refactors the codebase so that the state of both timers is
maintained by a state machine object.  The actual timing is in turn done
by a pair of timer objects.  The GUI elements are now just that: fairly
thin widgets that draw what the model says to, and send signals when
they're clicked.  Next up, portrait mode!

Incidentally, while doing this, it was very easy to add idle inhibition
when the timers are running, so I did. ^.^