Add portrait mode
With the new state machine implemented, it's now easy to add more
buttons to the interface.  This commit takes advantage of that, adding a
new portrait orientation with an AdwSqueezer.  I'm not aware of any
outstanding issues with it, but I wouldn't put it past my tipsy self to
have written a bug or two.