Add transparent headerbars to timer screen
Previously, we didn't have a headerbar.  We just wrapped the whole
window in a window handle, so you could drag anywhere to move the
window.  That caused problems however, especially on touchscreens, where
a small motion while pressing one of the buttons caused the press to be
ignored.  This is the crux of issue #3.

This commit fixes #3 by removing the window handle, and adding
transparent headerbars to the top of both landscape and portrait layouts
of the timer screen.  So now users can drag the window by the top, but
can't click the buttons there.  Still a weird compromise, because the
ideal behavior is hard to do with GTK.  But this should be a much more
practical tradeoff.