Split state setter into separate methods
The aim of this refactor is to get the increment code separated from the
rest of the state machine code, so each kind of increment can have its
own state machine subclass.  This commit does the first step of this,
splitting the state machine's transition function into a separate
method for each destination state.