Set a11y labels for timer buttons
Previously, the timer buttons had the same a11y labels as each other,
making it impossible to tell which player's button was selected when
using a screen reader.  This commit fixes that usability blunder by
labeling them appropriately as "player white/black timer", and changing
the labels when the button colors change.