Fix digit display in custom time seconds widget
Previously, the seconds display correctly showed two digits even when a
single-digit number was entered, but selecting another widget with such
a digit on the spinbutton would make it go back to not showing a leading
zero.  This commit fixes it by changing the text in the "output" signal,
rather than "value-changed".