Use WindowControls instead of a HeaderBar
Previously, we were using a GtkHeaderBar on the time selection screen,
mostly as a vestige from when the headerbar was actually visible on
that screen.  After switching to a flat headerbar and adding scrolling,
it interfered with clicks on visible widgets near the top of the screen,
which would be especially annoying on a phone in landscape mode where
vertical space is at a premium.

This commit fixes that papercut by removing the headerbar, and replacing
it with a pair of overlaid GtkWindowControls.  Now there's nothing
obstructing the top of the screen anymore, and clicks work as expected.